Move nordigen app behind to src
