+++ src/app-nordigen/banks/mbank-retail-brexplpw.js
@@ -1,31 +1,26 @@
 import { printIban, amountToInteger } from '../utils.js';
 
-/** @type {import('./bank.interface').IBank} */
+/** @type {import('./bank.interface.js').IBank} */
 export default {
-  institutionId: 'SANDBOXFINANCE_SFIN0000',
+  institutionId: 'MBANK_RETAIL_BREXPLPW',
 
   normalizeAccount(account) {
     return {
       account_id: account.id,
       institution: account.institution,
       mask: account.iban.slice(-4),
-      name: [account.name, printIban(account)].join(' '),
+      name: [account.displayName, printIban(account)].join(' '),
       official_name: account.product,
       type: 'checking'
     };
   },
 
   sortTransactions(transactions = []) {
-    return transactions.sort((a, b) => {
-      const [aTime, aSeq] = a.transactionId.split('-');
-      const [bTime, bSeq] = b.transactionId.split('-');
-
-      return Number(bTime) - Number(aTime) || Number(bSeq) - Number(aSeq);
-    });
+    return transactions.sort((a, b) => Number(b.transactionId) - Number(a.transactionId));
   },
 
   /**
-   *  For SANDBOXFINANCE_SFIN0000 we don't know what balance was
+   *  For MBANK_RETAIL_BREXPLPW we don't know what balance was
    *  after each transaction so we have to calculate it by getting
    *  current balance from the account and subtract all the transactions
    *
@@ -33,7 +28,7 @@ export default {
    *  it includes transaction placed during current day
    */
   calculateStartingBalance(sortedTransactions = [], balances = []) {
-    const currentBalance = balances.find((balance) => 'interimAvailable' === balance.balanceType);
+    const currentBalance = balances.find((balance) => 'interimBooked' === balance.balanceType);
 
     return sortedTransactions.reduce((total, trans) => {
       return total - amountToInteger(trans.transactionAmount.amount);
