+++ src/app-nordigen/banks/sandboxfinance-sfin0000.js
@@ -1,26 +1,31 @@
 import { printIban, amountToInteger } from '../utils.js';
 
-/** @type {import('./bank.interface').IBank} */
+/** @type {import('./bank.interface.js').IBank} */
 export default {
-  institutionId: 'MBANK_RETAIL_BREXPLPW',
+  institutionId: 'SANDBOXFINANCE_SFIN0000',
 
   normalizeAccount(account) {
     return {
       account_id: account.id,
       institution: account.institution,
       mask: account.iban.slice(-4),
-      name: [account.displayName, printIban(account)].join(' '),
+      name: [account.name, printIban(account)].join(' '),
       official_name: account.product,
       type: 'checking'
     };
   },
 
   sortTransactions(transactions = []) {
-    return transactions.sort((a, b) => Number(b.transactionId) - Number(a.transactionId));
+    return transactions.sort((a, b) => {
+      const [aTime, aSeq] = a.transactionId.split('-');
+      const [bTime, bSeq] = b.transactionId.split('-');
+
+      return Number(bTime) - Number(aTime) || Number(bSeq) - Number(aSeq);
+    });
   },
 
   /**
-   *  For MBANK_RETAIL_BREXPLPW we don't know what balance was
+   *  For SANDBOXFINANCE_SFIN0000 we don't know what balance was
    *  after each transaction so we have to calculate it by getting
    *  current balance from the account and subtract all the transactions
    *
@@ -28,7 +33,7 @@ export default {
    *  it includes transaction placed during current day
    */
   calculateStartingBalance(sortedTransactions = [], balances = []) {
-    const currentBalance = balances.find((balance) => 'interimBooked' === balance.balanceType);
+    const currentBalance = balances.find((balance) => 'interimAvailable' === balance.balanceType);
 
     return sortedTransactions.reduce((total, trans) => {
       return total - amountToInteger(trans.transactionAmount.amount);
