+++ src/app-nordigen/banks/tests/mbank-retail-brexplpw.spec.js
@@ -1,8 +1,8 @@
-import MbankRetailBrexplpw from '../mbank-retail-brexplpw';
+import MbankRetailBrexplpw from '../mbank-retail-brexplpw.js';
 
 describe('MbankRetailBrexplpw', () => {
   describe('#normalizeAccount', () => {
-    /** @type {import('../../nordigen.types').DetailedAccountWithInstitution} */
+    /** @type {import('../../nordigen.types.js').DetailedAccountWithInstitution} */
     const accountRaw = {
       iban: 'PL00000000000000000987654321',
       currency: 'PLN',
@@ -98,7 +98,7 @@ describe('MbankRetailBrexplpw', () => {
   });
 
   describe('#countStartingBalance', () => {
-    /** @type {import('../../nordigen-node.types').Balance[]} */
+    /** @type {import('../../nordigen-node.types.js').Balance[]} */
     const balances = [
       {
         balanceAmount: { amount: '1000.00', currency: 'PLN' },
