+++ src/app-nordigen/nordigen-node.types.ts
