+++ src/app-nordigen/nordigen.types.ts
@@ -4,7 +4,7 @@ import {
   Institution,
   Transactions,
   Balance
-} from './nordigen-node.types';
+} from './nordigen-node.types.js';
 
 export type DetailedAccount = Omit<NordigenAccountDetails, 'status'> & NordigenAccountMetadata;
 export type DetailedAccountWithInstitution = DetailedAccount & { institution: Institution };
